+++ pcs/usage.py
@@ -20,6 +20,7 @@ def full_usage():
     out += strip_extras(pcsd([],False))
     out += strip_extras(host([],False))
     out += strip_extras(alert([], False))
+    out += strip_extras(client([], False))
     print(out.strip())
     print("Examples:\n" + examples.replace(" \ ",""))
 
@@ -113,6 +114,7 @@ def generate_completion_tree_from_usage():
     tree["node"] = generate_tree(node([], False))
     tree["alert"] = generate_tree(alert([], False))
     tree["booth"] = generate_tree(booth([], False))
+    tree["client"] = generate_tree(client([], False))
     return tree
 
 def generate_tree(usage_txt):
@@ -176,6 +178,7 @@ Commands:
     host        Manage hosts known to pcs/pcsd.
     node        Manage cluster nodes.
     alert       Manage pacemaker alerts.
+    client      Manage pcsd client configuration.
 """
 # Advanced usage to possibly add later
 #  --corosync_conf=<corosync file> Specify alternative corosync.conf file
@@ -1782,11 +1785,29 @@ Commands:
         return output
 
 
+def client(args=[], pout=True):
+    output = """
+Usage: pcs client <command>
+Manage pcsd client configuration.
+
+Commands:
+    local-auth [<pcsd-port>] [-u <username>] [-p <password>]
+        Authenticate current user to local pcsd. This is requiered to run some
+        pcs commands which may require permissions of root user such as 'pcs
+        cluster start'.
+"""
+    if pout:
+        print(sub_usage(args, output))
+    else:
+        return output
+
+
 def show(main_usage_name, rest_usage_names):
     usage_map = {
         "acl": acl,
         "alert": alert,
         "booth": booth,
+        "client": client,
         "cluster": cluster,
         "config": config,
         "constraint": constraint,